backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==24.3
django-cors-headers==4.3.1
# 10.1.0 often fails to build on Windows + newer Python (no wheel → broken sdist). Use a release with wheels.
Pillow>=10.4.0,<12
python-decouple==3.8
python-dotenv==1.0.0
requests==2.31.0
cryptography==41.0.7
PyJWT==2.8.0
# 2.9.9 has no cp313 Windows wheel → pip builds from source and needs pg_config. 2.9.10+ ships wheels.
psycopg2-binary==2.9.10
gunicorn==21.2.0
whitenoise==6.6.0